refactor(spotify-service): add utility functions for string normalization
