+++ srcs/backend/services/content-service/src/services/spotify.service.ts
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { getRedisClient } from '../lib/redis.js';
+import { clearString, normalizeString } from '../utils/utils.js';
 
 interface AccessToken {
     access_token: string;
@@ -20,6 +21,18 @@ const MAX_LIMIT_FETCH = 10;
 
 let tokenFetchPromise: Promise<string | null> | null = null;
 
+/**
+ * @brief Fetches a new Spotify API access token using Client Credentials flow.
+ * 
+ * @details Makes an HTTP POST request to Spotify's token endpoint with the configured
+ * client credentials. The returned token is automatically cached in Redis with a TTL
+ * of (expires_in - 300) seconds to account for expiration buffer.
+ * 
+ * @return A promise that resolves to the access token string on success
+ * @throws Error if the HTTP response status is not 200 or if the request fails
+ * 
+ * @see getSpotifyToken()
+ */
 async function fetchSpotifyToken(): Promise<string | null> {
     const response = await axios.post<AccessToken>(
         'https://accounts.spotify.com/api/token',
@@ -30,13 +43,14 @@ async function fetchSpotifyToken(): Promise<string | null> {
         }).toString(),
         {
             headers: {
-                'Content-Type': 'application/x-www-form-urlencoded'
+                'Content-Type': 'pplication/x-www-form-urlencoded'
             }
         }
     )
 
     if (response.status !== 200) {
-        throw new Error('Couldn\'t obtain an access token');
+        console.error('Couldn\'t obtain an access token');
+        throw new Error();
     }
     console.log("Spotify API access token was retrieved");
 
@@ -51,12 +65,40 @@ async function fetchSpotifyToken(): Promise<string | null> {
     return access_token;
 }
 
+/**
+ * @brief Retrieves a Spotify API access token, with caching and request deduplication.
+ * 
+ * This function manages Spotify access token retrieval using the Client Credentials flow.
+ * It implements a two-level caching strategy:
+ * - Redis cache for persistent token storage across service instances
+ * - In-memory promise lock to prevent duplicate concurrent token requests
+ * 
+ * @details
+ * The function first attempts to retrieve a cached token from Redis. If a valid cached
+ * token exists, it is returned immediately. If no cache hit occurs and a token fetch is
+ * already in progress (indicated by the module-level @c tokenFetchPromise), that existing
+ * promise is returned instead of initiating a duplicate request. Otherwise, a new token
+ * fetch is initiated via @ref fetchSpotifyToken().
+ * 
+ * The @c tokenFetchPromise is reset in the finally block to ensure the in-memory lock
+ * is released regardless of success or failure.
+ * 
+ * @return A promise that resolves to:
+ *         - A string containing the access token if successful
+ *         - @c null if token retrieval fails or an error occurs
+ * 
+ * @see fetchSpotifyToken()
+ * @see getRedisClient()
+ * 
+ * @note The Spotify token is cached in Redis with a TTL of (expires_in - 300) seconds
+ *       to ensure tokens are refreshed before actual expiration.
+ */
 export async function getSpotifyToken(): Promise<string | null> {
     const redis = getRedisClient();
 
     const cachedToken = await redis.get('spotify_token');
     if (cachedToken) {
-        console.log(cachedToken);
+        // console.log(cachedToken);
         return cachedToken;
     }
 
@@ -74,112 +116,18 @@ export async function getSpotifyToken(): Promise<string | null> {
     }
 }
 
-// let tokenPromise: Promise<string> | null = null;
-// let tokenExpirationTime: number = 0;
-// let isFetching: boolean = false;
-
 /**
- * Retrieves an access token from Spotify using the Client Credentials flow.
+ * @brief Fetches tracks from Spotify API with pagination support.
+ * 
+ * @param term The search query term to send to Spotify
+ * @param offset The pagination offset (starting result index)
+ * @param token The Spotify API bearer token for authentication
+ * 
+ * @return A promise that resolves to the Axios response object containing track results
+ * 
+ * @note This function performs the actual API call without result deduplication.
+ *       Use @ref searchTracks() for a high-level interface with deduplication.
  */
-// async function getSpotifyToken(): Promise<string> {
-//     const currentTime = Date.now();
-
-//     // Return cached token if exists and is not within 5 minutes of expiring
-//     if (tokenPromise && (isFetching || currentTime < tokenExpirationTime)) {
-//         return tokenPromise;
-//     }
-
-//     isFetching = true;
-
-//     tokenPromise = (async () => {
-//         try {
-//             const response = await axios.post<AccessToken>(
-//                 'https://accounts.spotify.com/api/token',
-//                 new URLSearchParams({
-//                     grant_type: 'client_credentials',
-//                     client_id: CLIENT_ID || '',
-//                     client_secret: CLIENT_SECRET || ''
-//                 }).toString(),
-//                 {
-//                     headers: {
-//                         'Content-Type': 'application/x-www-form-urlencoded'
-//                     }
-//                 }
-//             );
-
-//             if (response.status !== 200) {
-//                 throw new Error('Couldn\'t obtain an access token');
-//             }
-
-//             console.log("Spotify API access token was retrieved successfully");
-
-//             const { access_token, expires_in } = response.data;
-
-//             // Calculate expiration time in ms, minus a 300-second (5-minute) buffer
-//             tokenExpirationTime = Date.now() + (expires_in - 300) * 1000;
-
-//             return access_token;
-//         } catch (error) {
-//             tokenPromise = null;
-//             tokenExpirationTime = 0;
-//             isFetching = false;
-//             throw error;
-//         }
-//     })();
-
-//     return tokenPromise;
-// }
-
-function stripBrackets(input: string, opening: string): string {
-    let result: string = "";
-    let depth: number = 0;
-    let closing: string = '';
-
-    if (input.normalize())
-
-        if (opening === '[') closing = ']';
-        else if (opening === '(') closing = ')';
-
-    for (let i = 0; i < input.length; i++) {
-        const char = input[i];
-
-        if (char === opening) {
-            depth++;
-        } else if (char === closing) {
-            if (depth > 0) depth--;
-        } else if (depth === 0) {
-            // Only append characters when not inside any brackets
-            result += char;
-        }
-    }
-
-    return result;
-}
-
-function normalizeString(str: string): string {
-    let result: string = '';
-
-    result = stripBrackets(str, '(');
-    result = stripBrackets(result, '[');
-
-    return result
-        .toLowerCase()
-        .replace(/\s+[-–—]\s+.*/g, '') // Removes all after -, – or —
-        .replace(/\s+/g, ''); // Removes all whitespace
-}
-
-function clearString(str: string): string {
-    let result: string = '';
-
-    result = stripBrackets(str, '(');
-    result = stripBrackets(result, '[');
-
-    return result
-        .replace(/\s+[-–—]\s+.*/g, '') // Removes all after -, – or —
-        .replace(/\s\s+/g, ' ') // Replace more than 1 whitespace with space
-        .trim();
-}
-
 async function fetchTracks(term: string, offset: number, token: string | null) {
     const response = axios.get('https://api.spotify.com/v1/search', {
         params: {
@@ -195,13 +143,33 @@ async function fetchTracks(term: string, offset: number, token: string | null) {
     })
 
     if ((await response).status === 429) {
-        console.log('hey');
         console.log((await response).headers);
     }
 
     return response;
 }
 
+/**
+ * @brief Searches Spotify for tracks matching the given term with deduplication.
+ * 
+ * @details Fetches tracks from two paginated result pages (20 total results) and
+ * returns up to 10 unique tracks. Uniqueness is determined by comparing normalized
+ * versions of track and artist names (case-insensitive, without special characters).
+ * 
+ * Uses @ref normalizeString() to create identifiers for deduplication and
+ * @ref clearString() to format display names.
+ * 
+ * @param term The search query to find on Spotify
+ * 
+ * @return A promise that resolves to an array of @c TrackData objects (max 10 unique tracks)
+ *         containing track name, artist name, and Spotify track ID
+ * 
+ * @see searchTracks()
+ * @see normalizeString()
+ * @see clearString()
+ * 
+ * @note Results are limited to the Spanish market (@c market: 'ES')
+ */
 export async function searchTracks(term: string): Promise<TrackData[]> {
     const token = await getSpotifyToken();
